Makes the initialization more smooth and less jumpy
We were hiding everything with an ng-if until we were fully initialized.
This means all DOM elements were not there yet, not taking any space.
As a result, when initialized all other elements of the page would
suddenly be pushed down by our DOM elements which looks quite bad.
We are also still removing the containing content group of Umbraco which
leads to some jumping. Perhaps we should also stop messing with that
default UI element and fix our own styling to fit into a group.