block.state.load -> block.state.loadEditor
More clearly conveys the meaning, became necessary after adding the
binding "isLoading" which indicates if the block should render its loading
view. state.loadEditor controls whether to render the inner nested content
editor or not. We lazily load it, i.e. it will not be loaded until a block
is opened for the first time.