Styling changes to render inside property group container.
This should probably have been the default, but by default we currently
actually hide the property group container so our editor can take even more
space. However, this does not work very well when other editors are put in
the same group as the container element is now completely gone so
properties appear to just float randomly on the page.