Added config option "Hide property group container"
When set to true the property group container that holds our editor
is hidden. Our editor will then take up all its space.
This defaults to true as we have been shipping our editor this way without
an option to turn this behavior off.

It is quite likely we will change the default to true in the near future,
as hiding the container messes up the default Umbraco UI and while it looks
fine when we are the only editor in that group, it does not anymore when
users add more editors to the same group.