Hide property group container: don't when inside some other property
In those cases the editor is technically not directly in a property group,
it seems unexpected the option would actually do anything in those cases so
to apply the principle of least surprise we will not hide the property
group container in those cases. We would be hiding the group container
of the parent property which is strange.