Removes unnecessary $timeout uses
This only delays logic until subsequent digest cycles and serves no purpose
as far as I can tell. The one around setPreviewScale() actually caused
noticable preview jumps in Edge and less so in Firefox (tested with
multiple Content Blocks editors on one page which probably played a role).
This may have been necessary in the past, but not anymore it seems. If this
ends up being absolutely necessary in some edge case we can just revert.