Handle DataTypeId / DataTypeKey the same way in C#/JS and add comments
In JS we would not use DataTypeId 0 because 0 is falsy in JS. In C# we checked
for NOT NULL, which is what the original intention was. JavaScript will now
perform the same check. This means if users accidently provide 0 as the
DataTypeId, this will immediately be clear in Umbraco because an exception will
be thrown when that data type cannot be found.

The direct reason for this change was issue #17 where it turned out a user had
provided both a DataTypeId (of 0) and a DataTypeKey. In the server side code
the validator would actually look at the Id (as it is not null), but client side
code would use the Key. Because everything seemed fine on the client it was hard
to identify the 0 Id as the cause of the issue.
This small change will make similar issues much easier to find later on.