Alter copy/paste behavior to not quit when header cannot be pasted
When a "Copy all blocks" action is performed on some ContentBlocks
instance which includes a header the paste action would not do anything
when executed in a ContentBlocks instance which also had a header
configured.

The idea was the copy/paste would be an all or nothing. Turns out this
isn't a great user experience either. Ideally, the user gets a choice to
overwrite the header or not.  Until then though, nudged by issue #20, we
do continue with pasting blocks rather than stopping everything.

If the user would like to overwrite the header anyway, they could simply
remove the existing header and then copy/paste the old header.