Fix jumpy blocks in block picker on hover in some situations
In certain situations the blocks would be very jumpy when hovering
over one block in the block picker.

This situation arises when the current amount of blocks can be displayed
in the picker almost exactly without a vertical scrollbar. When hovering
over a block, the highlight border that was drawn would slightly
increase the overall width/height and could cause the scrollbar to
appear, making all blocks jump.

This has now been fixed by using an outline that does not increase block
width/height.