Fix macro issues inside RTE (#31)
The property value of macros in RTEs were not stored correctly and as a
result rendering on the front-end was incorrect as well.

We have implemented a DataValueEditor on the server now that implements
FromEditor() + ToEditor() which in turn calls NestedContent's
FromEditor() + ToEditor() which do the heavy lifting for us.