Set value with correct culture & segment in temporary property for NC
When calling ToEditor() of Nested Content we pass in a temporary
property holding the value of 1 of our blocks. However, we always
created an invariant property (culture = null, segment = null) instead
of using the passed in culture and segment.

When calling ToEditor() from Nested Content with that property Nested
Content would then read an empty value in cases where culture or segment
are relevant such as multilingual doctypes.