Do not throw an exception when a definition cannot be found
This one is thrown in the backoffice and will prevent the whole editor
from loading. Aside from this specific case there is no good reason to
throw here rather than simply return null. It happens every now and then
that block definitions are deleted but still present in the property
data.