Fix preset blocks not initializing any property value
Presets would not load their property editor at all since they start
collapsed with loadEditor set to false. By default we lazily load
existing blocks on the page which greatly increases load time but
a new preset should be eagerly loaded instead to populate the property
value with the default empty value of that block.

Before this fix creating a new page with some presets and saving that
page would result in 0 blocks being saved because none of the blocks
actually set a property value in our editor.