Hide generic validation messages
Validation messages appear above the ContentBlocks editor and at the top
of every Content Block but do not provide any information other than
"this block has an error". We already show invalid blocks using a red
highlight so this is not needed and only confuses people.

Especially the generic validation message of the whole Content Blocks
property editor is confusing as it seems to point to the first block
but this message shows if any of the blocks is invalid.