Support .NET 5 + Umbraco 9
We use the multi targeting approach and target both .NET Framework (v8)
and .NET 5 (v9). A lot of the code remains unchanged and only .NET 5
or Umbraco 9 specific changes were made. The API has remained mostly
unchanged as well, only the rendering API was slightly changed.

In .NET 5 rendering partial views is recommended to be async so we adopt
this too. Rendering ContentBlocks is now done in this way in a view:

@inject IContentBlockRenderer Renderer

@await Html.RenderContentBlocks(Model.ContentBlocks, Renderer)

Alternatively, use the Renderer directly:
@await Renderer.Render(Model.ContentBlocks, Html.PartialAsync)