Fix Publish error on .NET 6
.NET 6 seems to have added more strict checks on the list of published
files and if it contains a duplicate output file it will fail.
We add the ContentBlocks content files to the list of published files to
ensure they are published but we added the files from the NuGet source
directory rather than the project directory. .NET 6 does not like it when
different files with the same destination in the published artifact exist
apparently so throws an error.
We now include the files from the project directory instead to avoid this.