Fix layout issue with Umbraco PR 11901
That PR adds a "top:50px" which for some reason pushes the toolbar down
when an RTE is used within a ContentBlock inside an Umbraco tab.
Our p-block__main has overflow: hidden but this does not seem to do anything
so we can remove it to fix this bugged interaction with that PR's change.