Fix overlay layout issue introduced by Umbraco 10.2.0
Due to a new CSS rule (contain: layout) our full screen picker overlays
were not rendering correctly anymore. We use a simple component to
render our overlays in another location of the DOM in order to not be
affected by this CSS rule that is only applied to the
.umb-property-editor container. Our overlays will now render outside of
it.